Moved some code and deleted a file no longer need
